+++ CHANGES.md
@@ -1,5 +1,24 @@
 # Calendar.js Change Log:
 
+### Version 0.6.0:
+- Added support to remove an event from a repeating series (asked in the confirmation message).
+- Holidays (Christmas Eve, Christmas Day, Boxing Day, New Year's Eve and New Year's Day) are now the default holidays.
+- Added navigation key support when full-screen mode is activated:
+  1. Left:  Moves to the previous month.
+  2. Right:  Moves to the next month.
+  3. Escape:  Closes full-screen mode.
+  4. Down:  Jumps to the current month.
+  6. F5:  Refreshes the current view.
+  7. Ctrl/Command + F:  Opens the search dialog.
+  8. Ctrl/Command + Left:  Moves to the previous year.
+  9. Ctrl/Command + Right:  Moves to the next year.
+- Added "created" support to events (set automatically if not available to the current date/time).
+- Added "organizerName" support to events (set automatically if not available to options > organizerName).
+- Added "organizerEmailAddress" support to events (set automatically if not available to options > organizerEmailAddress).
+- Added support to export events to the iCAL format.
+- Fixed events being in the wrong order when exported (to any format).
+- Added public methods "moveToPreviousYear()" and "moveToNextYear()" (with new "on" events).
+
 ### Version 0.5.2:
 - The text "All Day Event" is now "All-Day" (renamed all the settings to match).
 - Fixed a fault that allowed text to be selected in a disabled text field (date, time).